Fehler bei Laden mit Subs behoben (foreign key wurde falsch gesetzt)

+++ nicki-db/src/main/java/org/mgnl/nicki/db/context/derby/DerbyContext.java
@@ -72,7 +72,7 @@ public class DerbyContext
 	}
 
 	@Override
-	public int getSequenceNumber(String sequenceName) throws Exception {
+	public long getSequenceNumber(String sequenceName) throws Exception {
 		
 		return RandomUtils.nextInt();
 	}
